+++ gpu/command_buffer/service/shared_image/dxgi_swap_chain_image_representation.h
@@ -44,7 +44,7 @@ class GLTexturePassthroughDXGISwapChainBufferRepresentation
   ~GLTexturePassthroughDXGISwapChainBufferRepresentation() override;
 
   const scoped_refptr<gles2::TexturePassthrough>& GetTexturePassthrough(
-      int plane_index) override;
+      size_t plane_index) override;
 
  private:
   bool BeginAccess(GLenum mode) override;
